requirements.txt
dash
plotly
pandas
sqlalchemy
python-dotenvns